iter20 #fix 3. add positive scenario and move checker code to pkg
