+++ pkg/exitcheck/analyzer.go
