+++ pkg/exitcheck/testdata/src/a/main.go
