+++ pkg/exitcheck/testdata/src/b/main.go
@@ -1,0 +1,12 @@
+package main
+
+import (
+	"net/http"
+)
+
+func main() {
+	_, err := http.Get("http://example.com/")
+	if err != nil {
+		// handle error
+	}
+}
